Inserido o processo de Venda e ItemVenda
Descrição:
  - Implementado as Tabelas para realização da Venda no sistema, a "Venda" e a "ItemVenda".

  - Inserido processos para controle de "ItemVenda", onde não permite inserir o mesmo Produto duas
    vezes dentro da mesma Venda.

  - Inserido NumberUtilities para facilitar as validações com BigDecimal, validando se é maior que
    zero ou, maior que outro número.

+++ controleVenda/src/main/java/com/mycompany/controlevenda/model/Produto.java
@@ -1,5 +1,7 @@
 package com.mycompany.controlevenda.model;
 
+import java.math.BigDecimal;
+
 /**
  * Entidade Produto.
  *
@@ -20,12 +22,12 @@ public class Produto {
     /**
      * Preço do Produto.
      */
-    private double preco;
+    private BigDecimal preco;
 
     public Produto() {
     }
 
-    public Produto(String descricao, double preco) {
+    public Produto(String descricao, BigDecimal preco) {
         this.descricao = descricao;
         this.preco = preco;
     }
@@ -46,11 +48,11 @@ public class Produto {
         this.descricao = descricao;
     }
 
-    public double getPreco() {
+    public BigDecimal getPreco() {
         return preco;
     }
 
-    public void setPreco(double preco) {
+    public void setPreco(BigDecimal preco) {
         this.preco = preco;
     }
 }
